tests: split pre & post tests, and make work with new ast parser
